Files and login changes to organize between presentational and container components
