+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import LoginForm from './components/LoginForm'
+import LoginForm from './containers/Login'
 import {Provider} from 'react-redux'
 import store from './redux/store'
